version_1.3.1

+++ geometry-core/src/main/java/geometryLibrary/base/AreaCalculator.java
@@ -2,6 +2,6 @@ package geometryLibrary.base;
 
 public interface AreaCalculator{
 
-    double calculateArea();
+    long calculateArea();
 
 }
