+++ geometry-core/src/main/java/geometryLibrary/base/PerimeterCalculator.java
@@ -2,6 +2,6 @@ package geometryLibrary.base;
 
 public interface PerimeterCalculator{
 
-    double calculatePerimeter();
+    long calculatePerimeter();
 
 }
